third_party/WebKit/LayoutTests/external/wpt/fullscreen/interfaces-expected.txt

This is a testharness.js-based test.
PASS Test driver
FAIL Document interface: attribute fullscreenEnabled assert_equals: setter must be function for PutForwards, Replaceable, or non-readonly attributes expected "function" but got "undefined"
PASS Unscopable handled correctly for fullscreenEnabled property on Document
FAIL Document interface: attribute fullscreen assert_true: The prototype object must have a property "fullscreen" expected true got false
FAIL Unscopable handled correctly for fullscreen property on Document assert_equals: Document.prototype[Symbol.unscopables].fullscreen must exist expected "object" but got "undefined"
FAIL Document interface: operation exitFullscreen() assert_unreached: Throws "TypeError: Illegal invocation" instead of rejecting promise Reached unreachable code
PASS Unscopable handled correctly for exitFullscreen() on Document
PASS Document interface: attribute onfullscreenchange
PASS Unscopable handled correctly for onfullscreenchange property on Document
PASS Document interface: attribute onfullscreenerror
PASS Unscopable handled correctly for onfullscreenerror property on Document
FAIL Document interface: attribute fullscreenElement assert_equals: setter must be function for PutForwards, Replaceable, or non-readonly attributes expected "function" but got "undefined"
PASS Unscopable handled correctly for fullscreenElement property on Document
PASS Document interface: new Document must inherit property "fullscreenEnabled" with the proper type
FAIL Document interface: new Document must inherit property "fullscreen" with the proper type assert_inherits: property "fullscreen" not found in prototype chain
PASS Document interface: new Document must inherit property "exitFullscreen()" with the proper type
PASS Document interface: new Document must inherit property "onfullscreenchange" with the proper type
PASS Document interface: new Document must inherit property "onfullscreenerror" with the proper type
PASS Document interface: new Document must inherit property "fullscreenElement" with the proper type
FAIL ShadowRoot interface: attribute fullscreenElement assert_equals: setter must be function for PutForwards, Replaceable, or non-readonly attributes expected "function" but got "undefined"
PASS Unscopable handled correctly for fullscreenElement property on ShadowRoot
FAIL Element interface: operation requestFullscreen() assert_unreached: Throws "TypeError: Illegal invocation" instead of rejecting promise Reached unreachable code
PASS Unscopable handled correctly for requestFullscreen() on Element
PASS Element interface: attribute onfullscreenchange
PASS Unscopable handled correctly for onfullscreenchange property on Element
PASS Element interface: attribute onfullscreenerror
PASS Unscopable handled correctly for onfullscreenerror property on Element
PASS Element interface: document.createElementNS(null, "test") must inherit property "requestFullscreen()" with the proper type
PASS Element interface: document.createElementNS(null, "test") must inherit property "onfullscreenchange" with the proper type
PASS Element interface: document.createElementNS(null, "test") must inherit property "onfullscreenerror" with the proper type
Harness: the test ran to completion.
